+++ src/components/_global/BalTooltip/BalTooltip.vue
@@ -39,7 +39,8 @@ export default defineComponent({
     noPad: { type: Boolean, default: false },
     disabled: { type: Boolean, default: false },
     iconSize: { type: String, defailt: 'md' },
-    width: { type: String, default: '52' }
+    width: { type: String, default: '52' },
+    textCenter: { type: Boolean, default: false }
   },
   setup(props) {
     const activator = ref<HTMLElement>();
@@ -49,7 +50,8 @@ export default defineComponent({
     const tooltipClasses = computed(() => {
       return {
         'p-3': !props.noPad,
-        [`w-${props.width}`]: true
+        [`w-${props.width}`]: true,
+        'text-center': props.textCenter
       };
     });
